add tag control in config

+++ src/admin/Folders/Modal.svelte
@@ -1,11 +1,10 @@
 <script>
-  import { fade } from "svelte/transition";
   import TextAreaInput from "../Component/TextAreaInput.svelte";
   import Icons from "src/icons/Icons.svelte";
+  import Dialog from "../Disks/Dialog.svelte";
 
   export let file;
   export let modalType;
-  export let ref = null;
   export let deleteFromSys = null;
   export let hide;
   export let acept;
@@ -20,10 +19,10 @@
   };
 
   const submit = () => {
-    if (file.Name) {
+    if (tempFile.Ex) {
       file.Name = tempFile.Name + tempFile.Ex;
-      acept(file, deleteFromSys);
     }
+    acept(file, deleteFromSys);
   };
 
   const onKeyDown = (e) => {
@@ -39,53 +38,30 @@
 
   const onChanges = ({ target: { checked } }) => (deleteFromSys = checked);
 
-  if (file?.Name) loadTemp(file);
+  if (file?.Name && !modalType.Del) loadTemp(file);
 </script>
 
-<div class="modal-container">
-  <div class="modal card" transition:fade={{ duration: 200 }} on:keydown={onKeyDown} tabindex="-1">
-    <div class="modal-header">
-      <h4>{modalType.title}</h4>
-    </div>
-    <form bind:this={ref} action="#" on:submit|preventDefault={submit}>
-      <div class="modal-body">
-        {#if modalType.Del}
-          <p>Are you sure you want to remove <strong>{file.Name}</strong></p>
-          <div class="input-group">
-            <div class="input-group-prepend">
-              <label for="sysdel" class="input-group-text del-label"> Delete From System </label>
-            </div>
-            <label for="sysdel" class="form-control check-del">
-              <input id="sysdel" type="checkbox" on:change={onChanges} />
-              <Icons name={deleteFromSys ? "check" : "times"} />
-            </label>
-          </div>
-        {:else}
-          <TextAreaInput file={tempFile} key="Name" style="margin-bottom: 5px" rows="3" focus={true} />
-        {/if}
+<Dialog on:keydown={onKeyDown} confirm={submit} cancel={hide}>
+  <h4 slot="modal-header">{modalType.title}</h4>
+  <div id="fol-diag" slot="modal-body">
+    {#if modalType.Del}
+      <p>Are you sure you want to remove <strong>{file.Name}</strong></p>
+      <div class="input-group">
+        <div class="input-group-prepend">
+          <label for="sysdel" class="input-group-text del-label"> Delete From System </label>
+        </div>
+        <label for="sysdel" class="form-control check-del">
+          <input id="sysdel" type="checkbox" on:change={onChanges} />
+          <Icons name={deleteFromSys ? "check" : "times"} />
+        </label>
       </div>
-      <div class="error">{modalType.error || ""}</div>
-      <div class="modal-footer">
-        <button type="submit" class="btn">
-          {modalType.Del ? "Remove" : "Update"}
-        </button>
-        <button type="button" class="btn" on:click={hide}>Cancel</button>
-      </div>
-    </form>
+    {:else}
+      <TextAreaInput file={tempFile} key="Name" style="margin-bottom: 5px" rows="3" focus={true} />
+    {/if}
   </div>
-</div>
+</Dialog>
 
 <style>
-  .modal {
-    width: 400px;
-    outline: none;
-  }
-  .modal-container :global(.input-control) {
-    margin-bottom: 5px;
-  }
-  .modal-container :global(.input-label) {
-    width: 145px;
-  }
   .del-label {
     width: fit-content;
   }
@@ -100,10 +76,5 @@
   }
   label {
     text-align: center;
-  }
-  @media screen and (max-width: 450px) {
-    .modal {
-      width: 380px;
-    }
   }
 </style>
